+++ vexipy/__init__.py
@@ -1,5 +1,5 @@
 # ruff: noqa: F401
-from vexipy.component import Component, Subcomponent
+from vexipy.component import Product, Subcomponent
 from vexipy.document import Document
 from vexipy.statement import Statement
 from vexipy.status import StatusJustification, StatusLabel